+++ xarray/tests/test_dataarray.py
@@ -7416,3 +7416,54 @@ def test_clip(da):
     # Unclear whether we want this work, OK to adjust the test when we have decided.
     with pytest.raises(ValueError, match="arguments without labels along dimension"):
         result = da.clip(min=da.mean("x"), max=da.mean("a").isel(x=[0, 1]))
+
+
+pytest.mark.parametrize("keep", ["first", "last", False])
+def test_drop_duplicates(keep):
+    ds = xr.DataArray(
+        [0, 5, 6, 7], dims="time", coords={"time": [0, 0, 1, 2]}, name="test"
+    )
+
+    if keep == "first":
+        data = [0, 6, 7]
+        time = [0, 1, 2]
+    elif keep == "last":
+        data = [5, 6, 7]
+        time = [0, 1, 2]
+    else:
+        data = [6, 7]
+        time = [1, 2]
+
+    expected = xr.DataArray(data, dims="time", coords={"time": time}, name="test")
+    result = ds.drop_duplicates("time", keep=keep)
+    assert_equal(expected, result)
+
+
+@pytest.mark.parametrize("keep", ["first", "last", False])
+def test_drop_duplicates_multi_dim(keep):
+    base_data = np.stack([np.arange(0, 5) * i for i in np.arange(0, 5)])
+    ds = xr.DataArray(
+        base_data,
+        coords={"lat": [0, 1, 2, 2, 3], "lon": [0, 1, 3, 3, 4]},
+        dims=["lat", "lon"],
+        name="test",
+    )
+
+    if keep == "first":
+        data = base_data[[0, 1, 2, 4]][:, [0, 1, 2, 4]]
+        lat = [0, 1, 2, 3]
+        lon = [0, 1, 3, 4]
+    elif keep == "last":
+        data = base_data[[0, 1, 3, 4]][:, [0, 1, 3, 4]]
+        lat = [0, 1, 2, 3]
+        lon = [0, 1, 3, 4]
+    else:
+        data = base_data[[0, 1, 4]][:, [0, 1, 4]]
+        lat = [0, 1, 3]
+        lon = [0, 1, 4]
+
+    expected = xr.DataArray(
+        data, dims=["lat", "lon"], coords={"lat": lat, "lon": lon}, name="test"
+    )
+    result = ds.drop_duplicates(["lat", "lon"], keep=keep)
+    assert_equal(expected, result)